Rename and refactor image similarity search service;Remove unused scripts for test: 'hypha.py', 'find-similar-image.py', and 'test' Combine image vector ebmedding code into one code: ebmed-image-vectors.py
